+++ mostrar.h
@@ -1,8 +1,9 @@
+#include <stdint.h>
 /**
 * \brief Funcion que permite visualizar los registros de memoria
 * \param arreglo Dato de tipo long  el cual contiene los 12 registros de memoria
 * \return La funcion no tiene retorno
 */
-void visualizar(unsigned long int arreglo[12]);
+void visualizar(uint32_t arreglo[12]);
